Modification to the story
Added 'I quickly looked around the shed and saw a perfeclty cooked steak with chips, still piping hot, sitting on a plate on the table, as if by a God send. I almost cried with happiness. The tears got stronger as I made my way over to it and realised there was even a bottle of ketchup on the shelf above. Maybe if I was going to die... I would die happy. A strong crashing brought me out of my thoughts. I rolled my eyes in annoyance. I was not going to die in vain, especially if the zombies weren't even going to let me have a last meal in peace. I spotted an axe in the corner of the shed and snatched it up. It was go time.'